Fixed warnings

+++ AnotherOneBitesTheCrust/Map.cpp
@@ -11,7 +11,7 @@ Map::~Map(void)
 }
 
 Tile* Map::getTile(glm::vec3 position) {
-	int width = position.x / tileSize;
-	int depth = position.z / tileSize;
+	int width = (int)position.x / tileSize;
+	int depth = (int)position.z / tileSize;
 	return &tiles[width][depth];
 }
